# docs/admin/admin_swagger.yaml
basePath: /api
definitions:
...
  admin.OauthForm:
    properties:
      auto_register:
        type: boolean
      client_id:
        type: string
      client_secret:
        type: string
      id:
        type: integer
      issuer:
        type: string
      oauth_type:
        type: string
      op:
        type: string
      pkce_enable:
        type: boolean
      pkce_method:
        type: string
      scopes:
        type: string
    required:
    - client_id
    - client_secret
    - oauth_type
    type: object
...
  model.Oauth:
    properties:
      auto_register:
        description: RedirectUrl  string `json:"redirect_url"`
        type: boolean
      client_id:
        type: string
      client_secret:
        type: string
      created_at:
        type: string
      id:
        type: integer
      issuer:
        type: string
      oauth_type:
        type: string
      op:
        type: string
      pkce_enable:
        type: boolean
      pkce_method:
        type: string
      scopes:
        type: string
      updated_at:
        type: string
    type: object
  model.OauthList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Oauth'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: RustDesk API Server
  title: RustDesk API
  version: "1.0"
paths:
...
  /admin/oauth/create:
    post:
      consumes:
      - application/json
      description: Créer Oauth
      parameters:
      - description: Informations Oauth
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.OauthForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Oauth'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Créer Oauth
      tags:
      - Oauth
  /admin/oauth/delete:
    post:
      consumes:
      - application/json
      description: Supprimer Oauth
      parameters:
      - description: Informations Oauth
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.OauthForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Supprimer Oauth
      tags:
      - Oauth
  /admin/oauth/detail/{id}:
    get:
      consumes:
      - application/json
      description: Détails Oauth
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Oauth'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Détails Oauth
      tags:
      - Oauth
  /admin/oauth/list:
    get:
      consumes:
      - application/json
      description: Liste Oauth
      parameters:
      - description: Numéro de page
        in: query
        name: page
        type: integer
      - description: Taille de la page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.OauthList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Liste Oauth
      tags:
      - Oauth
  /admin/oauth/update:
    post:
      consumes:
      - application/json
      description: Modifier Oauth
      parameters:
      - description: Informations Oauth
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.OauthForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.OauthList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Modifier Oauth
      tags:
      - Oauth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"